# inventory-service/src/main/java/com/example/inventory_service/entity/Product.java
package com.example.inventory_service.entity;

import org.springframework.data.annotation.Id;
import org.springframework.data.relational.core.mapping.Column;
import org.springframework.data.relational.core.mapping.Table;

import java.time.LocalDateTime;
import java.util.UUID;

/**
 * Producto con stock simple (suposición del PRD: stock por producto, sin variantes).
 */
@Table("products")
public class Product {

    @Id
    @Column("product_id")
    private UUID productId;

    @Column("name")
    private String name;

    @Column("description")
    private String description;

    @Column("stock")
    private int stock;

    @Column("updated_at")
    private LocalDateTime updatedAt;

    public Product() {
    }

    public UUID getProductId() {
        return productId;
    }

    public void setProductId(UUID productId) {
        this.productId = productId;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public int getStock() {
        return stock;
    }

    public void setStock(int stock) {
        this.stock = stock;
    }

    public LocalDateTime getUpdatedAt() {
        return updatedAt;
    }

    public void setUpdatedAt(LocalDateTime updatedAt) {
        this.updatedAt = updatedAt;
    }
}